[copier] Move Copiers into sub directory
Note that the CopyOpBatcher doesn't keep track of array block indices
(dirty_ablocks) anymore. It was replaced by iterating through the
cleaned data blocks. See commits 79de213b and 84931913 for the details.
