+++ src/lib.rs
@@ -11,6 +11,7 @@ extern crate quickcheck_macros;
 pub mod cache;
 pub mod checksum;
 pub mod commands;
+pub mod copier;
 pub mod dump_utils;
 pub mod era;
 pub mod file_utils;
