+++ tests/cache_writeback.rs
@@ -10,9 +10,9 @@ use std::sync::{Arc, Mutex};
 
 use thinp::cache::mapping::*;
 use thinp::cache::superblock::*;
+use thinp::copier::test_utils::*;
 use thinp::file_utils::create_sized_file;
 use thinp::io_engine::buffer::Buffer;
-use thinp::io_engine::test_utils::*;
 use thinp::io_engine::{self, SyncIoEngine, PAGE_SIZE, SECTOR_SHIFT};
 use thinp::pdata::array::{self, *};
 use thinp::pdata::array_walker::*;
